update column type from double to decimal

+++ db/migrate/20180509034202_create_news_feeds.rb
@@ -3,8 +3,8 @@ class CreateNewsFeeds < ActiveRecord::Migration[5.2]
     create_table :news_feeds do |t|
       t.string :name
       t.text :description
-      t.double :lat
-      t.double :lng
+      t.decimal :lat
+      t.decimal :lng
       t.references :user, foreign_key: true
       t.boolean :is_private
       t.float :rating
